The lineariser is checked in for archival purposes.
The main problem that needs to be solved is that the
arguments of the linear processes do not have the right
form.

Besides that there are many other problems and bugs that
must be tracked down and repaired. 

The lineariser now works on .gs files, as illustrated
by the following example of its use, and its somewhat 
disappointing result:

jfg@antilope> linearise -regular ../examples/abp.gs
Fail to generate unique term of sort SortId("D")
Generation of constant terms has only partly been implemented





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@240 21252fe5-60e8-0310-8ff0-c1ae1001594e