The compiling rewriter has substantially been changed to avoid iterative transformation of terms in the linear process format making use of explicit application symbols, and the internal format where terms have the form appl#i(n,t1,t2,...,ti-1).
Before applying the compiling rewriter terms have to be translated
by gsToRewriteFormat. Terms that are returned must be translated
back by gsFromRewriteFormat. The translation functions are many times
more expensive than rewriting itself. In order to use the speed
of the rewriter it is strongly advised to avoid the use of the 
translation functions.

As it stands, the gsinstantiate must be adapted to use 
the compiling rewriter (gsinstantiate uses the interpreting
rewriter as its default).

The following need to be done:
1) substitution for variables in gsinstantiate must be done via
   the functions in gssubstitute.
2) the rewrite functions in libgsrewrite.c need to be cleaned up.
   It appears to me that only gsRewriteTerm and gsRewriteTerms (for
   lists) are needed. Verbose variants of the form 
   gsRewriteInternalWithSubsts, gsRewriteInternals, gsRewriteInternal
   seem to clutter the interface.
3) The function FindSolutions need to be adapted to work on the
   internal rewriting format. Muck seems to have done something, but
   I have no clue what has happened.
4) In the compiling rewriter the JITTY strategy must be applied.
   Currently, the innermost strategy is being applied.
   This can also be used to remove unused declarations.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@427 21252fe5-60e8-0310-8ff0-c1ae1001594e