mcrl22lpe contained an error where the time tag of an action was forgotten. This has been resolved.
mcrl22lpe now applies sum elimination when trying to resolve communications when
a data variable and a value must be synchronized. The value is substituted for the
variable, instead of adding a sum and an equality constraint.

There appears to be a serious problem with linearizing the alternating bit
protocol, as some summands appear to disappear. I will look into this with
highest priority.

Jan Friso



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@522 21252fe5-60e8-0310-8ff0-c1ae1001594e