Added data equations for subset-or-equal, proper subset, set union, set difference, set intersection and set complement. The following operations are not implemented: - Set2Bag because it belongs to the implementation of the corresponding Bag   sort, which is bigger - Set size because it seems impossible to me to implement this
Note that the addition of the above equations required the definition of lambda
expressions denoting logical negation, conjunction, disjunction and
implication. In order to simplify combinations of these functions, extra rules
need to be formulated on these operations. Theoretically this should not be
necessary because functions can be simplified using extensionality.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@572 21252fe5-60e8-0310-8ff0-c1ae1001594e