Previous commit was to hasty; Aad pointed out that some binaries had ended up in the repository.
- this is an attempt to remove those binaries

Improvements to build process:

- the debug versions of the boost libraries no longer build
- cleaning of trunk/source/boost is now done properly
- solved the toolset prefix mystery in boost library names
- configure and makefile deal with the included Boost libraries in a more sane fashion

Run the configure script with --with-boost=/your/own/boost/prefix to use your own version of boost to link.
Otherwise configure will resort to the included version.

A (manual) standard boost install should best be made with:

  bjam -sTOOLS=gcc --layout=system install

substitute gcc with the name of a supported some compiler tooset.

Remaining issue: boost/Makefile gives a hint of portability but I'm afraid it is no more than that.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@626 21252fe5-60e8-0310-8ff0-c1ae1001594e