Added mark states and mark deadlocks functionality.
The tool can now highlight clusters that contain deadlocks. Also, users can
define rules regarding the values of state parameters (e.g. "x is not an element
of {0,2,4}"). The user can choose whether all or any of the defined rules should
be satisfied by the states. A cluster will then be highlighted if any of its
states satisfies all/any of the rules. 

The functionality is a first version; any ideas for extensions or other
suggestions are welcome.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1163 21252fe5-60e8-0310-8ff0-c1ae1001594e