Changes: - generalised the lexer such that it can read multiple streams, by means of the   function parse_streams(std::vector<std::istream*> &streams); - removed the mcrl2de parser that was added yesterday - added library libparser:   + function parse_specification replaces the function mCRL2Parse from     mcrl2lexer.h   + function parse_data_expression replaces the function mCRL2ParseDE from     mcrlde_lexer.h.
Note that the new parse_data_expression function is implemented by faking the
parser to be a specification: a data expression DE is fed to the parser as the
string "init a@("DE");". On the one hand, this is a diry hack, but on the other
hand, it removes duplication of code and it forms the basis of a much more
general solution that will be added asap.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1320 21252fe5-60e8-0310-8ff0-c1ae1001594e