Updated the mcrl2 specification of the clobber game.
Added a specification of the domineering game. This game
has two players that in turn put a domino tile on a
checkers board. One does it horizontally, one vertically.
The player that cannot put a piece first looses the
game.

Furthermore added a description of the hex game. This
game, invented by Piet Hein, when not robbing silverfloats,
has two players, where one has to make a red connection
from left to right, and the other must make blue connection
from bottom to top by putting alternatingly red and blue
pieces on the board. Interesting property of this game
is that the who begins has a winning strategy, but the
strategy is unknown.

You see, I spend my time in a useful way, here in Reijkjavik.
Jan Friso



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1481 21252fe5-60e8-0310-8ff0-c1ae1001594e