A new version of ltsgraph is made. It is still quite clear that it does not behave as desired, but its behaviour is much more predictable now.
The major change is to remove the reference to nodes in edges
via an elaborate numbering scheme, by simply referring to the
nodes directly. This allows to draw the edges using the coordinates
of the nodes, which means that it is much more unlikely to see
loose edges or disappearing edges. Also the internal lookup and
data maintanance is much more straightforward.

The class Label has been removed. Labels are now stored in nodes
and edges. The files Label.h and Label.cpp should still be removed
somehow, Jeroen kun jij dit doen).

Furthermore, there is a new implementation of the force attraction
algorithm. Looking at them movement of the states on the screen
it should be obvious that there is still something wrong, but at
least the behaviour is quite consistent. In my opinion sufficiently
consistent to warrant checking this in.

It is fun to look at how this tool attempts to optimize, due to
the "live" interface of Didier.

Jan Friso



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1498 21252fe5-60e8-0310-8ff0-c1ae1001594e