Extended ATerm library interface with some additional protection functions. The standard interface has the following new functions:
  typedef void (*ATermProtFunc)();

  void ATaddProtectFunction(ATermProtFunc f);
  void ATremoveProtectFunction(ATermProtFunc f);
  void ATmarkTerm(ATerm *atp);
  void ATmarkArray(ATerm *start, int size);

The first two functions are to (un)register special functions that are
called during garbage collection such that they can mark additional ATerms
that need to be protected. This is done by calling ATmarkTerm or ATmarkArray
on an ATerm or array of ATerms, resp. (Just like you would otherwise protect
them with ATprotect or ATprotectArray.)

Besides this, there is an additional C++ interface in atermpp.h that allows
registration of objects that contain ATerms.

  class IProtectedATerm
  {
    public:
      virtual void ATprotectTerms() = 0;
  };

  void ATprotectProtectedATerm(IProtectedATerm *i);
  void ATunprotectProtectedATerm(IProtectedATerm *i);

Objects with ATerms can now implement the IProtectedATerm interface and
register it with the ATerm library for protection. Just as with the C
interface, the garbage collector will call the ATprotectTerms function for
all registered IProtectATerms, which in turn can use the ATmark* functions
to save the life of ATerms.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1507 21252fe5-60e8-0310-8ff0-c1ae1001594e