Implemented branch rotation for all visualization styles.
The default value of 111 degrees is computed by dividing 360 by the golden
section. This angle results from the study of phyllotaxis (how leaves on the
stem of a tree are arranged) and should produce pictures in which few branches
overlap. It has been taken from the botanical visualizations of Ernst Kleiberg,
Huub van de Wetering and Jack van Wijk.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1609 21252fe5-60e8-0310-8ff0-c1ae1001594e