*) introduced the class lpe::action_label *) renamed LPE::actions() and specification::actions() to action_labels(), and    updated some files according to this name change *) added an example for searching all action labels in an LPE (in lpe/example/find.cpp) *) added a check LPE::has_proper_action_labels to verify that all action labels of an    LPE l are included in l.action_labels().
N.B. This check is not yet added as a soundness check, since it turns out that most
example LPE's that I have violate this property.

I'd like to have a discussion about the need for an LPE carrying this redundant list
of action labels around. It doesn't seem right to store it inside an LPE.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2242 21252fe5-60e8-0310-8ff0-c1ae1001594e