*) Updated the LPE library according to the recent changes in the internal data format. Now both LPE and pbes use a data_specification internally.
*) All examples and tests of the LPE library now use inline specifications.

*) Made pp a global pretty print function for terms, to make the usage consistent.
So in all cases it is now:

Term t;
std::cout << pp(t);

*) Removed the member action_labels() from LPE (it was a mistake).
The files lpeconstelm.cpp and lpeparelm.cpp contained some calls to
lpe.action_labels().size(). These calls have been replaced by
spec.action_labels().size().

N.B. It seems awkward to me that these sizes are being used, since
they only give an upper bound for the number of action labels that
are present in the specification.

*) The files regfrmtrans.cpp and lin_std.cpp contained dynamically sized
C-arrays, which are not portable C++. Added patches for that to make them
compile with Visual C++. This means that the lineariser is now working
with Visual C++ too.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2295 21252fe5-60e8-0310-8ff0-c1ae1001594e