Changed the way in which the documentation is built.
Previously, every XML file was stand-alone, meaning it had its own header and
needed to be processed separately. Man-pages were built one at a time.

Now, there is a single master XML file (called index.xml) which includes all the
other XML files. By processing this single file, all man-pages are built in one
pass. This speeds up the process of building the documentation *DRAMATICALLY*,
because the required XSL-stylesheets and DTD-files (including their
dependencies) no longer have to be downloaded and processed for every XML file
over and over again. This is now done only once at the beginning.

Also, by simply passing a different stylesheet, the documentation can be converted
to HTML instead of MAN, resulting in a nice website. I still have to tweak this
to make the result look even nicer.

Another thing is that the master file (index.xml) contains LOTS of entities,
which have simply been collected from all other XML files. This list should be
cleaned up and compressed, which is rather tedious for a single person to do.

TO ALL AUTHORS: Please check your XML files and remove any entities that are
redundant, not necessary or not useful.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2583 21252fe5-60e8-0310-8ff0-c1ae1001594e