The lineariser has been adapted such that it will not generate ultimate delay constraints for processes that do not refer to time. Note however that there is a snag. A process of the form
   c -> p

implicitly contains time. I.e. it is interpreted as

   c -> p <> delta@0

In particular a process

   false -> p

is equal to delta@0.

Hence, when writing a specification it is often necessary
to add an <>delta (else delta) to avoid time constraints
to be generated in the parallel composition. I adapted most
examples (but not all) to contain the else delta option.

It is also possible to use the -D (add deltas) flag when
linearizing. This option will add a true -> delta summand
to each process, which suppresses time constraints. However,
this is not very elegant.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3241 21252fe5-60e8-0310-8ff0-c1ae1001594e