Changes:  - added --with-cvc= option to specify the prefix path to where CVC3 is    installed
 - updated the prover library to use cvc3 (executable) and cvc3 library instead
   of cvc lite

 - replaced references to cvc lite with CVC 3 in tools and the prover library
 
    note: CVC lite is deprecated and CVC3 is it's successor

People involved in the development of the data library please try installing
CVC and try to build the toolset. A lot of methods that were in
utilities::expression_info were removed but were still being used by the prover
(part that uses CVC)...  I have found some alternatives for these methods in
lps::data_expr, but not all of the methods seem to have an equivalent.

I hope that the functionality in sort_info.{h,cpp} and expr_info.{h,cpp} in the
utility can be put into the data library. Then the prover can be made whole
again and the files mentioned above can finally be removed from the repository.

Once again I want to ask everyone to take a look in the utilities library to
see if there are other things there that actually belong somewhere else.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3546 21252fe5-60e8-0310-8ff0-c1ae1001594e