Moved information tabs at the top of the right-hand panel to a separate dialog. This dialog can be accessed via the menu: File->Information.
The trigger for making this change was a bug noted by Jeroen van der Wulp: if
state parameter values became too large, the right-hand panel would resize
during simulation, leaving virtually no space for the visualization.

This commit resolves this bug.

The idea is that the remaining tabs on the right-hand panel (Mark and
Simulation) will also be "converted" to dialogs. The panel will then disappear,
giving all available space to the visualization.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3870 21252fe5-60e8-0310-8ff0-c1ae1001594e