The script can now generate the websites for a specific list of libraries only, instead of all libraries. Generating all websites takes a (relatively) long time which becomes annoying when writing documentation for just a single library.
To use this feature, pass the -l option to the script with a comma-separated
list of library IDs as the option-argument. The list of library IDs can be
requested using the -i option. Help is available via -h.

Again: note that the script should be called from the trunk directory only! I
have added a check for this, so it aborts with an error if it is called from a
different location.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3959 21252fe5-60e8-0310-8ff0-c1ae1001594e