Added the following feature to lpsconstelm:
Logging the value changes for the process parameters 
when performing the cycles. The values are stored in 
in a CSV file format separated by tabs. To use 
this feature, the short option "-c" or the long option "--csv", 
with a mandatory file need to be provided. This file is used to
store the logged data.    



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4205 21252fe5-60e8-0310-8ff0-c1ae1001594e