Substantially improved the performance of lpsrealelm by removing the calculation of the xi variables outside the calculation of zeta variables (this only happens when generating the summands, which is now done at the end of the program).
When the real numbers can only be approximated, due to a limited number
of iterations, may transitions are now generated properly.

Cleaned up messages.

What needs to be done is to rewrite Fourier-Motzkin, Gauss-elimination,
remove_redundant_inequalities and in particular is_inconsistent, because
these have now become the bottleneck, and do not work on any but the 
most trivial inequalities. This will be my next project.

Furthermore, a general code cleanup is needed.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5503 21252fe5-60e8-0310-8ff0-c1ae1001594e