Reimplemented all linear programming routines, and grouped them in a separate file, linear_inequalities.h, to be moved to the data library.
This appears to allow lpsrealelm to properly deal with
all linear inequalities. Unfortunately, the performance
does not substantially increase. Also, applying lpsuntime
and lpsrealelm to most variants of fischer's protocol
does not appear to terminate, because more and more
complex inequality constraints are being generated.
The precise reason for this is not understood by me.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5536 21252fe5-60e8-0310-8ff0-c1ae1001594e