Adapted gc.c in the ATerm library because its kwadratic garbage collection algorithms caused massive delays when memory usage is large. When 25GB was used, only 1/500 of the time was usefully used when generation the state space of the peg_solitaire example.
Currently, badly chosen moments to garbage collect by the aterm library
is still causing a 75% loss in useful performance, but for the moment
I will not try to improve this.

Erik Scheffers has incorporated the suggested changes also in the
ATerm library, but there is a real risk that our ATerm library
diverges from the official standard. 

If other examples show a notable performance increase I would
like to know that (esp. those from Yaroslav).



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5953 21252fe5-60e8-0310-8ff0-c1ae1001594e