Took care that the tools
   lysa2mcrl2
   mcrl2pp 
   garageframe.o
   
compile. They should work correctly, but this has not 
carefully been checked.

The only tool that does not compile now is mcrl2i.
This should have been resolved later today.

Thanks to Jeroen & Jeroen.

There are currently major problems with the
addition of standard datatypes in lps's and .pbes's
and the nextstate generator. These problems are
causing quite a number of tools to now work properly.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6052 21252fe5-60e8-0310-8ff0-c1ae1001594e