Made the generation of the parsers a bit more robust by replacing occurrences of
  #include "<prefix>parser.hpp"

in the generated <prefix>parser.cpp by

  #include "<proper_path>/<prefix>parser.hpp"


This resolves the issue that not all includes to files in the core library
were accessible using a single include directive (libraries/core/include).

BTW The location of the generated headers for the FSM and Dot parsers is
still "libraries/lts/include/mcrl2". Probably it is better to use
"libraries/lts/include/mcrl2/lts/detail".


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6129 21252fe5-60e8-0310-8ff0-c1ae1001594e