Completely revised the lineariser. It now uses the new data library, the lps library, does not have global variables, and does not refer to the gs and AT libraries, except for a small number of cases.
The lineariser has been tested on many major examples, 
and now appears to work. Given the number of mistakes
that appeared after the translation work was done,
it is likely that not all mistakes have been removed.

The new lineariser works completely as before, except
that in the comm command communication must always lead
to an action (i.e. comm({a|b->c},...) is ok, but
comm({a|b},...) and comm({a|b->tau},...) are now not
accepted anymore.

The names of parameters and sum variables have not been
preserved, nor has the sequence of process and sum paramaters.

The code of the lineariser has been reduced with well
over 2000 lines (to appr. 7000 lines) and it is likely
that its length can be reduced substantially in the
near future.

Revising the lineariser was a lot of work, and the
result is not really what one could consider as 
finalised.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6403 21252fe5-60e8-0310-8ff0-c1ae1001594e