Extended the linearisation test with lots of examples that caused problems when translating the linearizer to the new data library. Took care that the examples were run with different flag settings.
Currently, this test fails for two problems. The first one is that
there is a problem with alpha conversion. Yaroslav has been notified about this.

The second problem is that the new data library does not properly deal
with recursive types of the form:

sort T = struct  list(Nat,tail:T) | end;

The type of tail becomes: 

tail:struct  list(Nat,tail:T) | end  -> T

which means that the if tail is applied to an object of
some type, a different type results.

These problems must both be resolved.....




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6476 21252fe5-60e8-0310-8ff0-c1ae1001594e