The lineariser now applies constant elimination while linearising.
This can have a dramatic (kwadratic) effect on the number of summands
when processes communicate based on their identities, which are given as a parameter. This is typically
encountered in distributed protocols where each process is identical
and only differs in its address.

For constant elimination the standard library routine provided
in the lps library is used.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6904 21252fe5-60e8-0310-8ff0-c1ae1001594e