Ditch virtual destructor of binding_aware_traverser. This is thing that caused runtime failures in the windows MSVC build.
This could be a bug in the MSVC compiler, but at least it seems that strange things occur when classes that occur in more complex inheritance patters have virtual destructors. Possibly this is also related to the absence of other virtual functions, but this is currently not fully understood.

Fix found after an afternoon of painstaking debugging by Jeroen v.d. Wulp and Jeroen Keiren

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7032 21252fe5-60e8-0310-8ff0-c1ae1001594e