Frank noted that state spaces were generated for transitions of which the conditions did not rewrite to true or false, e.g.
  (head([])==3)-> a.X

would perform an a step. This turned out to be due to the fact that the enumerator
can operate in two modes, namely where elements are generated for conditions as long
as the condition is true, and alternatively as long as the condition is not false. The
first is required for state space generation, the second for for instance lpssuminst.
Unfortunately, it appears to me that the code taking care which one of the two is chosen
is unfinished. I made an adaption in the next_state generator, that could resolve the problem,
and which hopefully does not have side effects.

It turned out that enumeration and state space generation had ad-hoc error handling mechanisms.
These have been replaced by ordinary exceptions.

Three tests have been adapted. Two because they assumed that operators in the rewriters would
be ordered, which is not the case anymore. One because it was wrong.

One of the effects is that some tools, such as the simulators will not be very tolerant
anymore for under specified specifications, or specifications summing over reals.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7191 21252fe5-60e8-0310-8ff0-c1ae1001594e