Dealt with several problems in the lts library, in particular bug report #650 and #670.
In bug report 650 Krause showed that ltsconvert
would crash reading a .lts file. This was due
to a fixed size buffer in the .svc library that
was too small.

In #670 it was indicated that the --tau option
of ltsconvert and ltscompare were not working
properly. Now they can be used to hide any action
in an lts.

Checked by Jereon Keiren.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7384 21252fe5-60e8-0310-8ff0-c1ae1001594e