An example by Frank Stappers showed that rewrite rules of the form:
map f:Nat;
eqn f(1)=....

would not rewrite an expression of the form f(1).

This was due to the fact that the typechecker had been adapted to 
incorporate the data types of the new data library. An expression
1 of type nat would be translated by the type checker to Pos2Nat(1).
However, its normal form cNat(1). So, the rewrite rule
would be translated to:

eqn f(Pos2Nat(1))=....

whereas the term f(1) would be normalised to f(cNat(1)). Hence,
the rewrite rule does not apply to the term. This problem was already
noted by Aad Mathijssen, and he had solved this. A test case reflects
the problem. The typechecker has been adapted such that only
normal forms will be generated for concrete numbers of type real, int and nat.

Much credits goes to the test cases, which detected that creal has two
arguments, contrary to cint and cnat....

All testcases succeed again.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7743 21252fe5-60e8-0310-8ff0-c1ae1001594e