Added the possibility to generate counter traces when comparing two transitions systems with respect to bisimulation, branching bisimulation, divergence preserving branching bisimulation and strong and weak trace equivalence.
Moved the include file for traces to mcrl2/trace/trace.h instead of mcrl2/trace.h to 
be in line with the other libraries.

Adapted the trace library not to use malloced arrays, but to use vectors, in order
to achieve that it has a working copy constructor. 

Added a -c flag to ltscompare to generate counter examples in case two systems are not
equivalent. In libraries/lts/doc/tex/wiki-bisimulation-partitioner-notes.tex it is described
how these are generated, and what they mean. Essentially traces of the shape sigma.a are
generated, where sigma.a is a trace of one state, and from a non equivalent state sigma can
be done to a state where an a is not possible.

This was a major operation, so, there may be some undesired side effects. However, all 
tests succeed, and this has certainly be helpful to remove a few residual errors.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7774 21252fe5-60e8-0310-8ff0-c1ae1001594e