The representative generator did not yield correct terms for function sorts, and in its wake for sets and bags. This was noted by Mohammad Atif and Sjoerd Cranen.
Added an assert to guarantee that the representative generator always
yield terms of the requested sort.

Added an exception in case no proper sort is generated, instead of
using the default constructor for data terms.

Added a testcase to check whether the representative of the sort
Set(Real) is the empty set of proper type.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7809 21252fe5-60e8-0310-8ff0-c1ae1001594e