Removed an assert from nextstate_standard which would erroneously fire for an lps with zero summands. The assert was covered by other asserts anyhow.
Removed a very hard to find error in the jitty rewriter,
where in case a term

   f(x)

would be rewritten, where f is variable, to be replaced
by a complex expression, say g[1->true] (this is a function
update), then the new expression would consist of two
expressions, instead of four. One would get

    function_update(g)

which is syntactically nonsens, instead of 

    g[1->true](x)

which was desired.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7926 21252fe5-60e8-0310-8ff0-c1ae1001594e