pbes2bool did not handle lambda terms well, because the function "implement" was not applied to terms before rewriting. Implement replaces a lambda term by a fresh function symbol and adds a rewrite rule to the rewriting system.
In order to add implement a deprecated_pbes_rewriter has been added,
as a wrap around for the ordinary rewriter made by Wieger to obtain access
to the internals of the rewriter. 

The existence of the deprecated_pbes_rewriter allowed to remove 
the get_internal_rewriter() from Wieger's rewriter, which I had
added in the past to get access to the internals of the rewriter.
This entailed changes in exploration.cpp and nextstate_standard.cpp
as access to the internal functions of the rewriter had to be
obtained somehow, without the function get_internal_rewriter().

Furthermore, in bes_deprecated.h the rewriter is transformed
into a class variable, instead of a parameter. The same holds
for opt_precompile_pbes and internal_opt_store_as_tree which
have become const bool class parameters, instead of parameters
that are being passed around.

Finally, two asserts have been removed from linearise.cpp, 
because these where erroneous, as shown by an example of Sjoerd
Cranen.

Sjoerd also reviewed the code.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7980 21252fe5-60e8-0310-8ff0-c1ae1001594e