Repaired translation to internal terms by splitting data_specifications into those that are type checked, and those that are not. This should substantially reduce the use of sort normalisation, and even have a notable effect on the performance of the toolset... Details are below.
Matthias Raffelsieper showed (to my amazement) that it was possible
that in the presence of sort aliases not all natural numbers and 
set/bag enumerations would be replaced by their internal counterpart.
Concretely, consider the following example:

sort
  Bits = struct singleBit (bit: Bool)?isSingleBit | bitVector (bitVec:List(Bool))?isBitVector;
  t_sys_regset_fsm_state = Bits;
  t_timer_counter_fsm_state = Bits;

map
  repeat : Bool # Nat -> Bits;
  repeat_rec : Bool # Nat -> List(Bool);

var 
  b:Bool;
  n:Nat;
eqn 
  repeat(b,n) = if(n <= 1, singleBit(b), bitVector(repeat_rec(b,n)));

act a:Bits;
init a(repeat(true,32)).delta;

The rewrite rule was added in parsed format to the data specification (still
with the 1 as OpId("1"), i.e. before translation to internal format). 

At some point all equations were obtained to replace the 1 by its internal
format. For this purpose the normalised equations were taken, which differ
from the inserted equations. The normalised equation was translated and added,
but the old equation was not removed, because it was not the same as the normalised
equation. As a consequence there were now two equations one without the "1" and
one with the "1". As the "1" occurrs in the right hand side, it could appear in
terms when rewriting. If the "1" would occur at the left side, it would
not be usable by the rewrite system, and the only effect would be to slow the
rewriter down.

The solution to the problem was to let a data_specification to take care of
the translation of numbers and enumerated bags and sets by itself. This translation
is isolated in libraries/data/include/mcrl2/data/detail/translate_user_notation_to_internal_format.h.
The internal format conversion in libraries/data/include/mcrl2/data/detail/internal_format_conversion.h
of data specifications has been removed. 
  There was one problem, namely that when a data specification is not type checked, this translation
must not be done. The solution is that a data specification can be constructed before type checking,
but then no internal data structures are built, and no normalisation takes place. Only if it
is indicated that the data type is type checked, internal data structures are built, and 
translations performed. For this purpose the function 

declare_data_specification_to_be_type_checked()

is introduced, which can only and exactly be called once on a data specification.
If this function is not called, no mappings/constructors/equations/aliases of
the data type can be used. This may have to be changed, if the type checker is
replaced by a new one.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8005 21252fe5-60e8-0310-8ff0-c1ae1001594e