Replaced the implementation for lts2lps by a more direct and faster implementation. This required adapting the parser and typechecker for multi-actions. Furthermore, some additional asserts have been added to aterm_lists.
The previous implementation printed the lts as a textual .lps which
was subsequently parsed and type checked. The current implementation
parses the action labels and then generates the .lps directly.

It is now also possible to read a .lts and transform it directly
into an .lps without a separate data and action specification. Furthermore,
a .lps can now also be used as an action specification.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8026 21252fe5-60e8-0310-8ff0-c1ae1001594e