Fix compiling of command line interface with clang
The issue here is in overload resolution for operators << and >>. It seems that, according to the standard, a compiler should look in the std namespace and in the namespaces of the arguments to find the appropriate overloads. In this case, the definitions of the stream operators where in the mcrl2::utilities namespace, which was neither. Resolved the issue by moving the implementations to a more appropriate place, near the type definitions.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8439 21252fe5-60e8-0310-8ff0-c1ae1001594e