Fix compilation of pbespgsolve with clang
pbespgsolve uses the tr1 extensions if available. The specific extensions used however require -std=c++0x to be defined because of the use of variadic macros. Enabling c++0x support however leads to other errors. For the time being use the default implementation of sets and hashtables (std::set and std::map) in case clang is used.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8451 21252fe5-60e8-0310-8ff0-c1ae1001594e