Optimise translation used by besconvert.
Instead of one deadlock state, to which there is an edge from each other state, now a selfloop of the form self:block(...),op(...) is used to record block and boolean operand of an equation. As a result, not every state in each block is a bottom-state for the reduction algorithm. This induces a slight speedup in the reduction, but not as much as I had hoped for.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8522 21252fe5-60e8-0310-8ff0-c1ae1001594e