- Fixed lpsbinary, with the help of Jeroen.
It turned out that there is a difference in how the old replace_free_variables and the new substitute_free_variables functions treat assignments. Applying the substitution [x := y] to the assignment x = f(x) used to result in x = f(y), whereas the new version returns x = f(x). The fix for lpsbinary was to use substitute_variables instead of substitute_free_variables at one specific place in the code.

Note that there is a good reason to treat binding variables of assigments the same as binding variables of quantifiers and lambda expressions. For example when instantiating global variables, the function substitute_free_variables can be applied globally, without having to do something special for assigments. In the old situation this was not possible.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8793 21252fe5-60e8-0310-8ff0-c1ae1001594e