Partially reverted revision 8804 where based on experiments with Sjoerd the generation of case functions was changed.
The original situation was that case functions of the
shape 

  Case(e,v1,...,vn)

where vi are data expressions were simplified to v if all vi were
the same as v. In revision 8804 this simplification was also introduced
if all vi were equal to some value v, or global variables. This had 
an negative effect on the size of state spaces. E.g. the size of the abp
went from 74 to 78.

For the generation of the case functions in next state, the situation has been 
reverted, with the exception that if all vi are don't cares, a single
don't care is the result.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8952 21252fe5-60e8-0310-8ff0-c1ae1001594e